Added lite params, more tests
